+++ planet/api/models.py
@@ -13,15 +13,12 @@
 # limitations under the License.
 """Manage data for requests and responses."""
 
-import copy
 import datetime
-# import itertools
-import json
 import logging
 
 from ._fatomic import atomic_open
 from . import exceptions, utils
-from .. import specs
+
 
 CHUNK_SIZE = 32 * 1024
 
@@ -44,7 +41,7 @@ class Request(object):
         self.auth = auth
         self.params = params
 
-        body_type = body_type or Order
+        body_type = body_type or Body
         LOGGER.debug('Request body type: {}'.format(body_type))
         assert issubclass(body_type, Body)
         self.body_type = body_type
@@ -224,98 +221,6 @@ class JSON(Body):
         return self.response.json()
 
 
-class Order(JSON):
-    LINKS_KEY = '_links'
-    RESULTS_KEY = 'results'
-    LOCATION_KEY = 'location'
-
-    @property
-    def results(self):
-        '''Results for each item in order.'''
-        links = self.data[self.LINKS_KEY]
-        results = links.get(self.RESULTS_KEY, None)
-        return results
-
-    @property
-    def items_iter(self):
-        '''An iterator of the download locations for order results
-        The iterator yields the individual items in the order.
-
-        :return: iter of result download locations in order
-        '''
-        return (r[self.LOCATION_KEY] for r in self.results)
-
-    @property
-    def items(self):
-        '''Download locations for order results.
-
-        :return: list of result download locations in order
-        '''
-        return list(self.items_iter)
-
-    @property
-    def state(self):
-        return self.data['state']
-
-    @property
-    def id(self):
-        return self.data['id']
-
-
-class OrderDetailsException(Exception):
-    pass
-
-
-class OrderDetails(object):
-    '''Validating and preparing an order description for submission'''
-    BUNDLE_KEY = 'product_bundle'
-
-    def __init__(self, details):
-        self._data = copy.deepcopy(details)
-        self._validate_details()
-
-    @property
-    def products(self):
-        return self._data['products']
-
-    @property
-    def data(self):
-        '''The order details as a string representing json.'''
-        return json.dumps(self._data)
-
-    def _validate_details(self):
-        '''Try valiently to get details to match schema.
-
-        Checks that details match the schema and, where possible, change
-        the details to fit the schema (e.g. change capitalization')
-        '''
-        products = self.products
-        for p in products:
-            self._validate_bundle(p)
-            self._validate_item_type(p)
-
-    def _validate_bundle(self, product):
-        supported = specs.get_product_bundles()
-        self._substitute_supported(product, self.BUNDLE_KEY, supported)
-
-    def _validate_item_type(self, product):
-        key = 'item_type'
-        bundle = product[self.BUNDLE_KEY]
-        supported = specs.get_item_types(bundle)
-        self._substitute_supported(product, key, supported)
-
-    @staticmethod
-    def _substitute_supported(product, key, supported):
-        try:
-            matched_type = specs.get_match(product[key], supported)
-            LOGGER.debug('{}: {}'.format(key, matched_type))
-            product[key] = matched_type
-        except(StopIteration):
-            msg = '{} - \'{}\' not in {}'.format(
-                key, product[key], supported)
-            raise OrderDetailsException(msg)
-
-
 # class Orders(Paged):
 #     def _get_item_key(self):
 #         return 'orders'
@@ -325,8 +230,8 @@ class OrderDetails(object):
 #
 #     def _get_next_key(self):
 #         return 'next'
-
-
+#
+#
 # def _abstract():
 #     raise NotImplementedError("You need to override this function")
 #
